requirements.txt
Django==4.2
python-dotenv==1.0.0
python-telegram-bot==20.3
celery==5.3.1
redis==4.5.5
django-celery-results==2.5.1
Pillow==10.0.1